more work on sidebar
